Update to aas-core-meta, codegen, testgen 79314c6, 94399e1, e1087880
We update the development requirements to and re-generate everything
with:
* [aas-core-meta 79314c6],
* [aas-core-codegen 94399e1] and
* [aas-core3.0-testgen e1087880].

Notably, we fix the patterns for date and date-times with zone offset
`14:00` which previously allowed for a concatenation without a plus
sign.

In addition, we fix the test data for defects which were detected while
testing with the generated C++ SDK. This concerns the examples of
doubles which overflowed in C++, but where silently accepted otherwise.

[aas-core-meta 79314c6]: https://github.com/aas-core-works/aas-core-meta/commit/79314c6
[aas-core-codegen 94399e1]: https://github.com/aas-core-works/aas-core-codegen/commit/94399e1
[aas-core3.0-testgen e1087880]: https://github.com/aas-core-works/aas-core3.0-testgen/commit/e1087880