Replace test data first in the update script (#32)
We need to remove *all* the test data as the specification can introduce
breaking changes.

We also need to replace the test data before generating the test code as
breaking changes in specification cause the test code generation to fail
since it runs a couple of preliminary consistency checks against
the test data.