Add base64url encoding (#42)
The base64url encoding is heavily used on the server side, and most
users of the SDK use it for interactions with AAS servers.  Hence, we
include base64url encoding in the `common` module to make their life
easier even though it is not directly used in the SDK itself.

The change corresponds to [aas-core-codegen dd3d4b4].

Fixes #36.

[aas-core-codegen dd3d4b4]: https://github.com/aas-core-works/aas-core-codegen/commit/dd3d4b4d5a0842bdc920d12a933f76889dda4d13